Move dialog to lib/widgets
