+++ lib/routes/initialize_satscard/initialize_satscard_page.dart
@@ -7,11 +7,11 @@ import 'package:breez/bloc/satscard/satscard_actions.dart';
 import 'package:breez/bloc/satscard/satscard_bloc.dart';
 import 'package:breez/bloc/satscard/satscard_model.dart';
 import 'package:breez/bloc/satscard/satscard_op_status.dart';
-import 'package:breez/routes/initialize_satscard/satscard_operation_dialog.dart';
 import 'package:breez/theme_data.dart' as theme;
 import 'package:breez/utils/min_font_size.dart';
 import 'package:breez/widgets/back_button.dart' as backBtn;
 import 'package:breez/widgets/satscard/chain_code_field.dart';
+import 'package:breez/widgets/satscard/satscard_operation_dialog.dart';
 import 'package:breez/widgets/satscard/spend_code_field.dart';
 import 'package:breez/widgets/single_button_bottom_bar.dart';
 import 'package:breez_translations/breez_translations_locales.dart';
